Hash Generator separiert.
