+++ src/application/data/VocabObject.java
@@ -1,20 +1,19 @@
 package application.data;
 
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 public class VocabObject {
 	private String vocab;
 	private byte[] id;
 	private LevelObject level = new LevelObject();
+	private HashGenerator gen = new HashGenerator();
 
 	public VocabObject(String vocab) throws NoSuchAlgorithmException {
 		if(vocab == null || vocab.isEmpty()) {
 			throw new IllegalArgumentException();
 		}
 		this.vocab = vocab;
-		this.id = getSHA(vocab);
+		this.id = gen.getSHA(vocab);
 	}
 
 	/**
@@ -62,19 +61,4 @@ public class VocabObject {
 	public void levelUp() {
 		level.levelUp();
 	}
-	
-	/**
-	 * Generates hash for id
-	 * @param input
-	 * @return
-	 * @throws NoSuchAlgorithmException
-	 */
-	public byte[] getSHA(String input) throws NoSuchAlgorithmException {
-		if(input == null || input.isEmpty()) {
-			throw new IllegalArgumentException();
-		}
-		MessageDigest md = MessageDigest.getInstance("SHA-256");
-		return md.digest(input.getBytes(StandardCharsets.UTF_8));
-	}
-
 }
